refactor(composer): fold the persona pill into the model/effort picker
Retires the standalone PersonaPill in favour of a Persona flyout inside ModelEffortPicker (same side-flyout pattern as Effort, shown only when the fleet-wide persona switch is on). The picker's design moves into a new PersonaOptions component that owns the store I/O, so the two composer dropdowns collapse to one. useDismissable's doc comment updated to match.
